[#1003-FIX] new issues
